modified project architecture
