+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\react\\start\\src\\index.js":"1","D:\\react\\start\\src\\reportWebVitals.js":"2","D:\\react\\start\\src\\App.js":"3","D:\\react\\start\\src\\components\\counter.jsx":"4"},{"size":501,"mtime":1608104981298,"results":"5","hashOfConfig":"6"},{"size":362,"mtime":499162500000,"results":"7","hashOfConfig":"6"},{"size":599,"mtime":1608104293249,"results":"8","hashOfConfig":"6"},{"size":898,"mtime":1608103323847,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1125dkx",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"19","usedDeprecatedRules":"20"},"D:\\react\\start\\src\\index.js",[],["21","22"],"D:\\react\\start\\src\\reportWebVitals.js",[],"D:\\react\\start\\src\\App.js",[],"D:\\react\\start\\src\\components\\counter.jsx",["23"],"import React, { useState } from 'react';\r\nimport { Container, Row, Col} from \"react-bootstrap\";\r\nimport Button from 'react-bootstrap/Button';\r\n\r\nfunction Counter (props) {\r\n\t  \treturn ( \r\n\t\t<Container>\r\n\t\t\t<Row className=\"justify-content-md-center\">\r\n\t\t\t\t<Col md=\"auto\" className=\" text-center\">\r\n                <h1>First counter</h1>\r\n\t\t\t\t</Col>\r\n\t\t\t</Row>\r\n\t\t\t<Row className=\"justify-content-md-center\">\r\n\t\t\t\r\n\t\t\t<Col><Button variant=\"primary\" onClick={props.increament}>+</Button></Col>\r\n            <Col><h5>{props.value?`Your Count : ${props.value}`:\"Zero\"}</h5></Col>\r\n            <Col><Button variant=\"primary\" onClick={props.decrement} >-</Button></Col>\r\n\r\n\t\t\t</Row>\r\n\t\t\t<Row className=\"justify-content-md-center\">\r\n\t\t\t\t<Col md=\"auto\" > \r\n\t\t\t\t <Button variant=\"outline-danger\" onClick={props.set}>Reset</Button>\r\n\t\t\t\t</Col>\r\n\t\t\t</Row>\r\n\t\t</Container>\r\n\t );\r\n}\r\n \r\nexport default Counter;\r\n",["24","25"],{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","severity":1,"message":"31","line":1,"column":17,"nodeType":"32","messageId":"33","endLine":1,"endColumn":25},{"ruleId":"26","replacedBy":"34"},{"ruleId":"28","replacedBy":"35"},"no-native-reassign",["36"],"no-negated-in-lhs",["37"],"no-unused-vars","'useState' is defined but never used.","Identifier","unusedVar",["36"],["37"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\MSI\\Desktop\\Counter\\src\\index.js":"1","C:\\Users\\MSI\\Desktop\\Counter\\src\\reportWebVitals.js":"2","C:\\Users\\MSI\\Desktop\\Counter\\src\\components\\App\\App.js":"3","C:\\Users\\MSI\\Desktop\\Counter\\src\\components\\Containers\\Containers.js":"4"},{"size":533,"mtime":1608155886564,"results":"5","hashOfConfig":"6"},{"size":375,"mtime":1608155401052,"results":"7","hashOfConfig":"6"},{"size":675,"mtime":1608156205798,"results":"8","hashOfConfig":"6"},{"size":1020,"mtime":1608156261049,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"bei0az",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\MSI\\Desktop\\Counter\\src\\index.js",[],["19","20"],"C:\\Users\\MSI\\Desktop\\Counter\\src\\reportWebVitals.js",[],"C:\\Users\\MSI\\Desktop\\Counter\\src\\components\\App\\App.js",[],"C:\\Users\\MSI\\Desktop\\Counter\\src\\components\\Containers\\Containers.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
