+++ src/components/App/App.scss
@@ -1,0 +1,9 @@
+.App {
+  display: flex;
+  width: 100vw;
+  height: 100vh;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  color: blueviolet;
+}
